refactor(objectives): move Source screening contract into analysis
Why:
- Confirmed-Objective screening owned Source traversal and fallback, but its model judgment was still split across generic extractor, prompt, and schema modules.

What:
- Co-locate screening input bounds, prompt, response contract, Source-id accounting, repair, and model call with the existing analysis stage.

Impact:
- Relevance decisions, batch traversal, repaired and fallback provenance, downstream routing, persistence, and public contracts remain unchanged.

Verification:
- All 169 screening, workflow, and extractor tests and 6 collection-build integration tests passed.
